Login-system-django/requirement.txt
Django==4.1.5
djangorestframework==3.13.1
django-crispy-forms==1.14.0